requirements.txt
Django==4.1.1
Pillow==9.2.0
django-bootstrap5==22.1
django-phonenumber-field==7.0.0
phonenumberslite==8.12.57